Create tfrecords how-to (#14)
* conda environment exported with no builds

* jupyter notebook for training, testing, evaluating, and visualizing the hourglass pose estimation model

* jupyter notebook for training, testing, evaluating, and visualizing the hourglass pose estimation model

* Feature/hourglass evaluation (#13)

Refactoring and consolidating evaluation code.

* removing duplicate files

* Update evaluation.py

* adding AWS roles required for annotation jobs

* added how to create tfrecords in jupyter notebook

* Jupyter notebook now includes how to create tfrecords using the utilities in pose_annotation_tools
