+++ .gitignore
@@ -2,3 +2,4 @@
 *.ckpt
 .ipynb_checkpoints
 .idea
+hourglass_pose\__pycache__\*
