+++ hourglass_pose/training_input.py
