+++ hourglass_pose/unused/get_GT_from_tfrecords.py
