+++ pose_annotation_tools/AWS_roles/SageMakerGroundTruthServerlessRecipeDeployment.json
@@ -1,0 +1,107 @@
+{
+    "Version": "2012-10-17",
+    "Statement": [
+        {
+            "Sid": "VisualEditor0",
+            "Effect": "Allow",
+            "Action": [
+                "iam:UpdateAssumeRolePolicy",
+                "iam:GetPolicyVersion",
+                "iam:GetAccountPasswordPolicy",
+                "iam:ListRoleTags",
+                "iam:ListServerCertificates",
+                "iam:GenerateServiceLastAccessedDetails",
+                "iam:ListServiceSpecificCredentials",
+                "iam:ListSigningCertificates",
+                "iam:ListVirtualMFADevices",
+                "iam:ListSSHPublicKeys",
+                "iam:SimulateCustomPolicy",
+                "iam:DetachRolePolicy",
+                "iam:SimulatePrincipalPolicy",
+                "iam:ListAttachedRolePolicies",
+                "iam:ListRolePolicies",
+                "iam:GetAccountAuthorizationDetails",
+                "iam:GetCredentialReport",
+                "iam:ListPolicies",
+                "iam:GetServerCertificate",
+                "iam:GetRole",
+                "iam:ListSAMLProviders",
+                "iam:GetPolicy",
+                "iam:GetAccessKeyLastUsed",
+                "iam:ListEntitiesForPolicy",
+                "iam:DeleteRole",
+                "iam:GetUserPolicy",
+                "iam:ListGroupsForUser",
+                "iam:GetGroupPolicy",
+                "iam:GetOpenIDConnectProvider",
+                "iam:GetRolePolicy",
+                "iam:GetAccountSummary",
+                "iam:GenerateCredentialReport",
+                "iam:GetServiceLastAccessedDetailsWithEntities",
+                "iam:ListPoliciesGrantingServiceAccess",
+                "iam:ListMFADevices",
+                "iam:GetServiceLastAccessedDetails",
+                "iam:GetGroup",
+                "iam:GetContextKeysForPrincipalPolicy",
+                "iam:GetOrganizationsAccessReport",
+                "iam:GetServiceLinkedRoleDeletionStatus",
+                "iam:ListInstanceProfilesForRole",
+                "iam:PassRole",
+                "iam:GenerateOrganizationsAccessReport",
+                "iam:ListAttachedUserPolicies",
+                "iam:ListAttachedGroupPolicies",
+                "iam:GetSAMLProvider",
+                "iam:ListAccessKeys",
+                "iam:GetInstanceProfile",
+                "iam:ListGroupPolicies",
+                "iam:GetSSHPublicKey",
+                "iam:ListRoles",
+                "iam:ListUserPolicies",
+                "iam:ListInstanceProfiles",
+                "iam:GetContextKeysForCustomPolicy",
+                "iam:ListPolicyVersions",
+                "iam:ListOpenIDConnectProviders",
+                "iam:ListAccountAliases",
+                "iam:ListUsers",
+                "iam:GetUser",
+                "iam:ListGroups",
+                "iam:GetLoginProfile",
+                "iam:ListUserTags"
+            ],
+            "Resource": "*"
+        },
+        {
+            "Sid": "VisualEditor1",
+            "Effect": "Allow",
+            "Action": [
+                "lambda:CreateFunction",
+                "lambda:ListVersionsByFunction",
+                "lambda:GetLayerVersion",
+                "lambda:GetAccountSettings",
+                "lambda:GetFunctionConfiguration",
+                "lambda:GetLayerVersionPolicy",
+                "lambda:ListProvisionedConcurrencyConfigs",
+                "lambda:GetProvisionedConcurrencyConfig",
+                "lambda:ListTags",
+                "lambda:ListLayerVersions",
+                "lambda:ListLayers",
+                "lambda:DeleteFunction",
+                "lambda:ListCodeSigningConfigs",
+                "lambda:GetAlias",
+                "lambda:ListFunctions",
+                "lambda:GetEventSourceMapping",
+                "lambda:GetFunction",
+                "lambda:ListAliases",
+                "lambda:GetFunctionCodeSigningConfig",
+                "lambda:ListFunctionEventInvokeConfigs",
+                "lambda:ListFunctionsByCodeSigningConfig",
+                "lambda:GetFunctionConcurrency",
+                "lambda:GetFunctionEventInvokeConfig",
+                "lambda:ListEventSourceMappings",
+                "lambda:GetCodeSigningConfig",
+                "lambda:GetPolicy"
+            ],
+            "Resource": "*",
+        }
+    ]
+}
